refactor: :art: change folder location of stacks
