+++ lib/stacks/bastion-host-stack.ts
