+++ lib/stacks/compute-stack.ts
