+++ lib/stacks/endpoint-stack.ts
