+++ lib/stacks/network-stack.ts
